Keep vuart TEMPTY interrupt status flag asserted
The TEMPTY interrupt status flag is set when writing the TX FIFO, but
cleared when the interrupt is acknowledged. This causes a problem if
a write to the TX FIFO occurs in the guest's interrupt handler, because
the interrupt could be acknowledged after writing to the TX FIFO.

The vuart device immediately flushes all writes to the TX FIFO, so it
is always empty. This commit will keep the TEMPTY flag set
indefinitely to more closely emulate real hardware behavior.

Change-Id: I8896e902168705d66d45ff7464b06af108a4f752